Extract rules deploy + status ops into host/rules.ts
Move setRules/setFirestoreRules/setDatabaseRules/getActiveRules/getRulesStatus
and the firestoreRuleMessages + normalizeDatabaseRules helpers into
host/rules.ts behind isRulesOp/handleRulesOp. handleOp routes to it after the
db/lens preamble. Bodies verbatim; no behavior change.
