+++ packages/pyric-tools/src/serve/worker/host.ts
@@ -143,11 +143,11 @@ import {
   lensProvenance,
   sessionDb,
   lensRtdb,
-  ensureRtdb,
   opProvenance,
 } from './host/core.js';
 import { isFirestoreReadOp, handleFirestoreReadOp } from './host/firestore-reads.js';
 import { isFirestoreWriteOp, handleFirestoreWriteOp } from './host/firestore-writes.js';
+import { isRulesOp, handleRulesOp } from './host/rules.js';
 
 // Re-export so host.ts's public surface is unchanged after the decomposition.
 export { ensureAuth, portSession } from './host-auth.js';
@@ -326,36 +326,6 @@ function rtdbSnapToWire(snap: DataSnapshot): unknown {
   };
 }
 
-function normalizeDatabaseRules(source: unknown): { rules: Record<string, unknown> } | null {
-  if (source === null) return null;
-  if (typeof source === 'string') {
-    return JSON.parse(source) as { rules: Record<string, unknown> };
-  }
-  if (typeof source === 'object' && source !== null) {
-    return source as { rules: Record<string, unknown> };
-  }
-  throw new Error('RTDB rules must be a rules JSON object or JSON string.');
-}
-
-function firestoreRuleMessages(result: { warnings?: Array<{ severity?: string; message?: string }>; parseError?: { line?: number; column?: number; expected?: unknown; actual?: string } | null }) {
-  const messages: Array<{ severity: 'info' | 'warn' | 'error'; text: string; line?: number; column?: number }> = [];
-  if (result.parseError) {
-    messages.push({
-      severity: 'error',
-      text: `PARSE ERROR: expected ${String(result.parseError.expected ?? 'valid rules')}`,
-      line: result.parseError.line,
-      column: result.parseError.column,
-    });
-  }
-  for (const warning of result.warnings ?? []) {
-    messages.push({
-      severity: warning.severity === 'error' ? 'error' : warning.severity === 'warning' ? 'warn' : 'info',
-      text: String(warning.message ?? warning),
-    });
-  }
-  return messages;
-}
-
 async function handleOp(ctx: HostCtx, port: PortLike, msg: OpMessage): Promise<void> {
   // Explicit lens (Studio admin / as / app-session) → lensDb; no lens → the
   // PORT'S SESSION (#754), so app ops run as whoever this tab signed in as.
@@ -372,61 +342,9 @@ async function handleOp(ctx: HostCtx, port: PortLike, msg: OpMessage): Promise<v
   // computed above; peeled off to their family modules.
   if (isFirestoreReadOp(msg.method)) return handleFirestoreReadOp(ctx, port, msg, db);
   if (isFirestoreWriteOp(msg.method)) return handleFirestoreWriteOp(ctx, port, msg, db);
+  if (isRulesOp(msg.method)) return handleRulesOp(ctx, port, msg, db);
 
   switch (msg.method) {
-    case 'setRules':
-    case 'setFirestoreRules': {
-      try {
-        const result = sandboxOps.setRules(db, msg.source);
-        const messages = firestoreRuleMessages(result);
-        const okDeploy = !messages.some((m) => m.severity === 'error');
-        ctx.activeRules ??= {};
-        const previous = ctx.activeRules.firestore?.status === 'active'
-          ? ctx.activeRules.firestore.source
-          : ctx.activeRules.firestore?.lastKnownGood;
-        ctx.activeRules.firestore = {
-          source: okDeploy ? msg.source : ctx.activeRules.firestore?.source ?? msg.source,
-          updatedAt: Date.now(),
-          status: okDeploy ? 'active' : 'error',
-          messages,
-          ...(previous ? { lastKnownGood: previous } : {}),
-        };
-        ok(port, msg.id, { warnings: result.warnings, messages, ok: okDeploy });
-      } catch (e) { fail(port, msg.id, e); }
-      break;
-    }
-
-    case 'setDatabaseRules': {
-      try {
-        const db = ensureRtdb(ctx);
-        const rules = normalizeDatabaseRules(msg.source);
-        const previous = ctx.activeRules?.database?.status === 'active'
-          ? ctx.activeRules.database.source
-          : ctx.activeRules?.database?.lastKnownGood;
-        rtdbSandbox.setRules(db, rules);
-        ctx.activeRules ??= {};
-        ctx.activeRules.database = {
-          source: rules,
-          updatedAt: Date.now(),
-          status: 'active',
-          messages: [],
-          ...(previous ? { lastKnownGood: previous } : {}),
-        };
-        ok(port, msg.id, { ok: true, messages: [] });
-      } catch (e) { fail(port, msg.id, e); }
-      break;
-    }
-
-    case 'getActiveRules': {
-      ok(port, msg.id, msg.service ? ctx.activeRules?.[msg.service] ?? null : ctx.activeRules ?? {});
-      break;
-    }
-
-    case 'getRulesStatus': {
-      ok(port, msg.id, msg.service ? ctx.activeRules?.[msg.service] ?? null : ctx.activeRules ?? {});
-      break;
-    }
-
     case 'admin.getDocument': {
       try {
         ok(port, msg.id, ctx.sandbox.admin.getDocument(msg.path));
